api/requirements.txt
fastapi==0.101.0
uvicorn==0.23.2
pydantic==2.1.1
SQLAlchemy==2.0.19
psycopg2==2.9.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
#python-dotenv==1.0.0
#python-decouple==3.8